Fix data race in send/recv (continue)
The previous commit f31c466 was not complete, as the tests failed.

I had to rewrite the way we perform `send` to make it safe again.

The issue this time was mainly caused by inconsistent update of the `lap` variable.

the writer, will update the `lap`, and then it will start writing the node, the reader can slip in the middle of these 2 events, and get an incorrect lap value.